added autorefresh
