+++ internal/ui/web/src/lib/components/ui/button-group/index.ts
@@ -1,0 +1,3 @@
+import Root from './button-group.svelte';
+
+export { Root, Root as ButtonGroup };
