+++ internal/ui/web/src/lib/domains/devices/components/device-table/device-columns.ts
