+++ internal/ui/web/src/lib/domains/devices/components/device-table/device-table-skeleton.svelte
@@ -1,0 +1,26 @@
+<script lang="ts">
+	import { Skeleton } from '$lib/components/ui/skeleton';
+	import * as Table from '$lib/components/ui/table';
+	import { deviceColumns } from './device-columns';
+</script>
+
+<Table.Root class="min-w-225">
+	<Table.Header>
+		<Table.Row class="hover:bg-transparent">
+			{#each deviceColumns as col, i (i)}
+				<Table.Head class="h-10 text-xs font-medium tracking-wide text-muted-foreground uppercase">
+					{col.header as string}
+				</Table.Head>
+			{/each}
+		</Table.Row>
+	</Table.Header>
+	<Table.Body>
+		{#each { length: 5 }, i (i)}
+			<Table.Row class="hover:bg-transparent">
+				{#each deviceColumns, j (j)}
+					<Table.Cell><Skeleton class="h-4 w-full" /></Table.Cell>
+				{/each}
+			</Table.Row>
+		{/each}
+	</Table.Body>
+</Table.Root>
