--- internal/ui/web/src/lib/domains/devices/services/device.ts
@@ -1,15 +1,0 @@
-import { api } from '../../../shared/services/api';
-import type { ApiResponse, ContextsResponse } from '../../../shared/types/api';
-
-/**
- * Context service for fetching Mir contexts from the API
- * Provides methods to retrieve available contexts and current context
- */
-export const deviceService = {
-	/**
-	 * Get all contexts from the API
-	 */
-	async getAll(): Promise<ApiResponse<ContextsResponse>> {
-		return api.get<ContextsResponse>('/v1/contexts');
-	}
-};
